requirements.txt
runway-python==0.0.74
numpy==1.16.3
pandas==0.24.2
scikit-learn==0.20.3
scipy==1.2.1
termcolor==1.1.0
faiss==1.5.3
